# libs/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al-Fath Berkarya API
servers:
  - url: /api
    description: Same-origin (production / single-container)
  - url: http://localhost:3000/api
    description: Local dev API server
tags:
  - name: ai
    description: AI completion
  - name: health
    description: Health check
  - name: members
    description: Community member profiles
  - name: matches
    description: Member matches
  - name: interests
    description: Shared interest vocabulary
  - name: karya
    description: Karya — the unit of work people build and collaborate around
  - name: feed
    description: Global activity feed and curated homepage picks
  - name: otp
    description: Email OTP verification

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /ai/stream:
    post:
      operationId: aiStream
      tags: [ai]
      summary: Stream a chat completion
      description: >
        Returns a `text/plain` chunked response. Consume with the `useStream`
        hook from `@myapp/ai/react` — not the generated client function.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteRequest"
      responses:
        "200":
          description: Streamed plain-text chunks
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/complete:
    post:
      operationId: aiComplete
      tags: [ai]
      summary: Complete a chat conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteRequest"
      responses:
        "200":
          description: AI response text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI provider not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /me:
    get:
      operationId: getMe
      tags: [members]
      summary: Get the current user's profile
      responses:
        "200":
          description: Profile if it exists, null otherwise
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/Member"
                  - type: "null"

  /profile:
    post:
      operationId: saveProfile
      tags: [members]
      summary: Create or update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all community members
      responses:
        "200":
          description: Array of member profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"

  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /matches:
    get:
      operationId: getMatches
      tags: [matches]
      summary: Get the current user's matches
      responses:
        "200":
          description: Array of matched member profiles with match reasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberMatch"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: saveMatches
      tags: [matches]
      summary: Save the current user's matches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveMatchesRequest"
      responses:
        "200":
          description: Matches saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /interests:
    get:
      operationId: listInterests
      tags: [interests]
      summary: List the shared interest vocabulary
      responses:
        "200":
          description: Array of interests, curated first then alphabetical
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interest"

  /karya:
    get:
      operationId: listKarya
      tags: [karya]
      summary: List all karya, reverse-chronological
      responses:
        "200":
          description: Array of karya with roster faces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KaryaListItem"
    post:
      operationId: createKarya
      tags: [karya]
      summary: Create a karya
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KaryaInput"
      responses:
        "200":
          description: Karya created — returns its id for redirect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KaryaCreated"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /karya/{id}:
    get:
      operationId: getKarya
      tags: [karya]
      summary: Get a karya by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Karya detail with roster, viewer membership, pending requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KaryaDetail"
        "404":
          description: Karya not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /karya/{id}/join:
    post:
      operationId: joinKarya
      tags: [karya]
      summary: Request to join a karya
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The viewer's resulting membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KaryaMembership"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Karya not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /karya/{id}/members/{userId}/approve:
    post:
      operationId: approveKaryaMember
      tags: [karya]
      summary: Approve a pending join request (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the karya owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Karya not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /karya/{id}/members/{userId}/decline:
    post:
      operationId: declineKaryaMember
      tags: [karya]
      summary: Decline a pending join request (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Request declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the karya owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Karya not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /karya/{id}/posts:
    get:
      operationId: getKaryaPosts
      tags: [karya]
      summary: List a karya's update stream, reverse-chronological
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of posts with author faces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "404":
          description: Karya not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPost
      tags: [karya]
      summary: Post an update to a karya (members only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "200":
          description: The created post with its author face
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Invalid kind or empty body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an approved member of this karya
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Karya not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /karya/{id}/feature:
    post:
      operationId: featureKarya
      tags: [karya]
      summary: Mark a karya as featured (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureInput"
      responses:
        "200":
          description: Karya featured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Karya not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unfeatureKarya
      tags: [karya]
      summary: Remove a karya from featured (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Karya unfeatured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Karya not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /feed:
    get:
      operationId: getFeed
      tags: [feed]
      summary: Reverse-chronological feed of recent posts and new karya
      responses:
        "200":
          description: Merged feed items, newest first (no ranking)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedItem"

  /featured:
    get:
      operationId: getFeatured
      tags: [feed]
      summary: Hand-curated "Top picked" karya for the homepage
      responses:
        "200":
          description: Featured karya, ordered by rank
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KaryaListItem"

  /otp/send:
    post:
      operationId: sendOtp
      tags: [otp]
      summary: Send a 6-digit OTP to the given student email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpSendRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid or non-student email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /otp/verify:
    post:
      operationId: verifyOtp
      tags: [otp]
      summary: Verify an OTP code and mark the email as verified
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        ok:
          type: boolean
        version:
          type: string
      required: [ok, version]

    Message:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required: [role, content]

    CompleteRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
          minItems: 1
      required: [messages]

    CompleteResponse:
      type: object
      properties:
        text:
          type: string
      required: [text]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]

    Member:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        handle:
          type: [string, "null"]
        bio:
          type: [string, "null"]
        interests:
          type: array
          items:
            type: string
        year:
          type: string
        major:
          type: string
        skills:
          type: array
          items:
            type: string
      required: [id, name, handle, bio, interests, year, major, skills]

    MemberMatch:
      allOf:
        - $ref: "#/components/schemas/Member"
        - type: object
          properties:
            reason:
              type: string
          required: [reason]

    ProfileInput:
      type: object
      properties:
        name:
          type: string
        handle:
          type: string
        bio:
          type: string
        interests:
          type: array
          items:
            type: string
        year:
          type: string
        major:
          type: string
        skills:
          type: array
          items:
            type: string
      required: [name, year, major, skills]

    Interest:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        curated:
          type: boolean
      required: [id, name, slug, curated]

    KaryaStage:
      type: string
      enum: [idea, validating, building, shipped, paused]

    RosterMember:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        handle:
          type: [string, "null"]
        image:
          type: [string, "null"]
      required: [id, name, handle, image]

    KaryaInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        stages:
          type: array
          items:
            $ref: "#/components/schemas/KaryaStage"
        interests:
          type: array
          items:
            type: string
      required: [title, description]

    KaryaCreated:
      type: object
      properties:
        id:
          type: string
      required: [id]

    Karya:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        stages:
          type: array
          items:
            $ref: "#/components/schemas/KaryaStage"
        interests:
          type: array
          items:
            type: string
      required: [id, title, description, stages, interests]

    KaryaMembership:
      type: object
      properties:
        role:
          type: string
        status:
          type: string
      required: [role, status]

    KaryaListItem:
      allOf:
        - $ref: "#/components/schemas/Karya"
        - type: object
          properties:
            roster:
              type: array
              items:
                $ref: "#/components/schemas/RosterMember"
            memberCount:
              type: integer
          required: [roster, memberCount]

    KaryaDetail:
      allOf:
        - $ref: "#/components/schemas/Karya"
        - type: object
          properties:
            createdBy:
              type: [string, "null"]
            roster:
              type: array
              items:
                $ref: "#/components/schemas/RosterMember"
            viewerMembership:
              anyOf:
                - $ref: "#/components/schemas/KaryaMembership"
                - type: "null"
            pendingRequests:
              type: array
              items:
                $ref: "#/components/schemas/RosterMember"
            featured:
              type: boolean
            viewerIsAdmin:
              type: boolean
          required:
            [
              createdBy,
              roster,
              viewerMembership,
              pendingRequests,
              featured,
              viewerIsAdmin,
            ]

    PostKind:
      type: string
      enum: [progress, challenge, achievement]

    PostInput:
      type: object
      properties:
        kind:
          $ref: "#/components/schemas/PostKind"
        body:
          type: string
          minLength: 1
      required: [kind, body]

    Post:
      type: object
      properties:
        id:
          type: string
        karyaId:
          type: string
        kind:
          $ref: "#/components/schemas/PostKind"
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/RosterMember"
      required: [id, karyaId, kind, body, createdAt, author]

    KaryaRef:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
      required: [id, title]

    FeedPostItem:
      allOf:
        - $ref: "#/components/schemas/Post"
        - type: object
          properties:
            type:
              type: string
              enum: [post]
            karya:
              $ref: "#/components/schemas/KaryaRef"
          required: [type, karya]

    FeedKaryaItem:
      allOf:
        - $ref: "#/components/schemas/KaryaListItem"
        - type: object
          properties:
            type:
              type: string
              enum: [karya]
            createdAt:
              type: string
              format: date-time
          required: [type, createdAt]

    FeedItem:
      oneOf:
        - $ref: "#/components/schemas/FeedPostItem"
        - $ref: "#/components/schemas/FeedKaryaItem"
      discriminator:
        propertyName: type
        mapping:
          post: "#/components/schemas/FeedPostItem"
          karya: "#/components/schemas/FeedKaryaItem"

    FeatureInput:
      type: object
      properties:
        rank:
          type: integer

    MatchEntry:
      type: object
      properties:
        memberId:
          type: string
        reason:
          type: string
      required: [memberId, reason]

    SaveMatchesRequest:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/MatchEntry"
      required: [matches]

    OtpSendRequest:
      type: object
      properties:
        email:
          type: string
      required: [email]

    OtpVerifyRequest:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
      required: [email, code]